% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msic.R
\name{fit_window}
\alias{fit_window}
\title{Find the best step arrangement in one window}
\usage{
fit_window(x, level_in, stats, n_total, rate_sum)
}
\arguments{
\item{x}{intensities of this window (reversed time).}

\item{level_in}{number of active fluorophores at the start of the window.}

\item{stats}{numeric vector `c(mB, vB, mF, vF)`: background / single-fluorophore
mean and variance.}

\item{n_total}{number of points in the whole (padded) trace.}

\item{rate_sum}{output of [rate_sum()].}
}
\value{
A list with `jumps` (as long as `x`, non-zero at steps) and `level`
  (number of active fluorophores at the end of the window).
}
\description{
Find the best step arrangement in one window
}
\keyword{internal}
