% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msic.R
\docType{data}
\name{JUMPS}
\alias{JUMPS}
\title{Allowed changes in fluorophore number per step}
\format{
An object of class \code{integer} of length 6.
}
\usage{
JUMPS
}
\description{
Allowed changes in fluorophore number per step
}
\keyword{internal}
