% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{priors}
\alias{priors}
\title{Priors for one trajectory}
\usage{
priors(t_last, mB, vB, mF, vF)
}
\arguments{
\item{t_last}{frame (0-based) of the last bleaching step (last fluorophore ->
background).}

\item{mB, vB}{mean and variance of the background (after the last step).}

\item{mF}{mean intensity added by one fluorophore.}

\item{vF}{variance of the single-fluorophore level.}
}
\value{
A list of class `photobleach_priors`.
}
\description{
What the user supplies for each trajectory, all in raw intensity units. Usually
created by [click_priors()] or read from a file with [read_priors()].
}
\examples{
p <- priors(t_last = 120, mB = 1000, vB = 62500, mF = 1500, vF = 185000)
}
